로컬 스토로지 api 분리
=> 이제 사용자가 클래스에서만 수정하면 됩니다.

- Cart, Comment, Product, QnA, Rivew , User 클래스의 유효성 체크는 무조건 true가 나옵니다.
- 사용 시에 데이터를 수정 하면됩니다.

- storageApi의 경우 클래스가 분리되면서 다소 옮겨졌습니다.
