+++ src/api/storageApi.js
@@ -6,393 +6,17 @@
 // enum(열거형 타입)을 만든다.
 // Result는 결과의 enum이고,
 // Class는 클래스명의 enum이다.
-const storageEnum = {
-  Result: {
-    Success: "success",
-    Failure: "failure",
-    Unknown: "unknown",
-  },
-
-  Class: {
-    Product: "product",
-    Cart: "cart",
-    User: "user",
-    Comment: "comment",
-    Review: "review",
-    QnA: "qna",
-  },
-};
-
-const getIdByClass = {
-  [storageEnum.Class.Product]: "productId",
-  [storageEnum.Class.Cart]: "cartId",
-  [storageEnum.Class.User]: "userId",
-  [storageEnum.Class.Comment]: "commentId",
-  [storageEnum.Class.Review]: "reviewId",
-  [storageEnum.Class.QnA]: "qnaId",
-};
-
-// Product클래스입니다.
-class Product {
-  // 생성자
-  constructor(productId, name, price, image) {
-    // this.productId = productId;
-    this.productId = Product.getNextId();
-    this.name = name;
-    this.price = price;
-    this.image = image;
-  }
-
-  // 데이터 체크
-  static validate(data) {
-    return (
-      (typeof data.productId === "number" || data.productId === "") &&
-      typeof data.name === "string" &&
-      typeof data.price === "number" &&
-      typeof data.image === "string"
-    );
-  }
-
-  // 저장소에서 pno의 값을 확인합니다.
-  static getNextId() {
-    const products =
-      JSON.parse(localStorage.getItem(storageEnum.Class.Product)) || [];
-
-    // 데이터에 문제가 있다면 1번을 부여합니다.
-    if (isValid(products) === false) return 1;
-
-    // 마지막 번호 + 1
-    const maxPno = Math.max(
-      ...products.map((product) => product.productId || 0)
-    );
-    return maxPno + 1;
-  }
-}
-
-// Cart클래스입니다.
-class Cart {
-  // 생성자
-  constructor(product, model, quantity, select) {
-    this.product = product;
-    this.model = model;
-    this.quantity = quantity;
-    this.select = select;
-  }
-
-  // 데이터 체크
-  static validate(data) {
-    return (
-      Product.validate(data.product) &&
-      typeof data.model === "string" &&
-      typeof data.quantity === "number" &&
-      Array.isArray(data.select)
-    );
-  }
-}
-
-// User 클래스 입니다.
-class User {
-  // 생성자 입니다.
-  // 다음 userId를 탐색하여 부여합니다.
-  constructor(userId, id, pwd) {
-    this.userId = isValid(userId) ? userId : User.getNextId();
-    this.id = id;
-    this.pwd = pwd;
-  }
-
-  // User.validate();
-  // 데이터 체크입니다.
-  // 모두 true로 판정이 나와야 성공입니다.
-  static validate(data) {
-    console.log("userData", data);
-    return (
-      // createEntitiy에서 pno를 ""로 넘기게 됩니다.
-      typeof data.id === "string" && typeof data.pwd === "number"
-    );
-  }
-
-  // 저장소에서 pno의 값을 확인합니다.
-  static getNextId() {
-    const users =
-      JSON.parse(localStorage.getItem(storageEnum.Class.User)) || [];
-
-    // 데이터에 문제가 있다면 1번을 부여합니다.
-    if (isValid(users) === false) return 1;
-
-    // 마지막 번호 + 1
-    const maxPno = Math.max(...users.map((user) => user.userId || 0));
-    return maxPno + 1;
-  }
-}
-
-class Comment {
-  constructor(user, commentId, content, date) {
-    this.user = user;
-    // this.commentId = isValid(commentId) ? commentId : Comment.getNextId();
-    this.commentId = Comment.getNextId();
-    this.content = content;
-    this.date = date;
-
-    console.log("commetId", this.commentId);
-  }
-
-  // Comment.validate();
-  // 데이터 체크입니다.
-  // 모두 true로 판정이 나와야 성공입니다.
-  static validate(data) {
-    console.log(User.validate(data.user));
-    console.log(typeof data.commentId === "number" || data.commentId === "");
-    console.log(typeof data.content === "string");
-    console.log(typeof data.date === "string");
-    return (
-      User.validate(data.user) &&
-      typeof data.commentId === "number" &&
-      // data.commentPno.trim() === "" &&
-      typeof data.content === "string" &&
-      typeof data.date === "string"
-    );
-  }
-
-  // 저장소에서 pno의 값을 확인합니다.
-  static getNextId() {
-    const comments =
-      JSON.parse(localStorage.getItem(storageEnum.Class.Comment)) || [];
-
-    // 데이터에 문제가 있다면 1번을 부여합니다.
-    if (isValid(comments) === false) return 1;
-
-    const maxPno = Math.max(
-      ...comments.map((comment) => comment.commentId || 0)
-    );
-
-    console.log("review pno", maxPno);
-    return maxPno + 1;
-  }
-}
-
-class Review {
-  constructor(reviewId, product, user, title, content, date, comments) {
-    // this.reviewId = isValid(reviewId) ? reviewId : Review.getNextId();
-    this.reviewId = Review.getNextId();
-    this.product = product;
-    this.user = user;
-    this.title = title;
-    this.content = content;
-    this.date = date;
-    this.comments = comments;
-  }
-
-  // Comment.validate();
-  // 데이터 체크입니다.
-  // 모두 true로 판정이 나와야 성공입니다.
-  static validate(data) {
-    console.log("test", data);
-    console.log(Product.validate(data.product));
-    console.log(User.validate(data.user));
-    console.log(typeof data.title === "string");
-    console.log(typeof data.content === "string");
-
-    return (
-      //  || data.reviewId.trim() === ""
-      // typeof data.reviewId === "number" &&
-      Product.validate(data.product) &&
-      User.validate(data.user) &&
-      typeof data.title === "string" &&
-      typeof data.content === "string"
-    );
-  }
-
-  // 저장소에서 pno의 값을 확인합니다.
-  static getNextId() {
-    const reviews =
-      JSON.parse(localStorage.getItem(storageEnum.Class.Review)) || [];
-
-    // 데이터에 문제가 있다면 1번을 부여합니다.
-    if (isValid(reviews) === false) return 1;
-
-    const maxPno = Math.max(...reviews.map((reviews) => reviews.reviewId || 0));
-    return maxPno + 1;
-  }
-}
-
-class QnA {
-  constructor(qnaId, user, title, content, date) {
-    this.qnaId = QnA.getNextId();
-    this.user = user;
-    this.title = title;
-    this.content = content;
-    this.date = date;
-  }
-
-  static validate(data) {
-    return true;
-  }
-
-  // 저장소에서 pno의 값을 확인합니다.
-  static getNextId() {
-    const qnas = JSON.parse(localStorage.getItem(storageEnum.Class.QnA)) || [];
-
-    // 데이터에 문제가 있다면 1번을 부여합니다.
-    if (isValid(qnas) === false) return 1;
-
-    const maxPno = Math.max(...qnas.map((qna) => qnas.qnaId || 0));
-    return maxPno + 1;
-  }
-}
-
-// Product , Cart , User 를 생성해줍니다.
-// classType은 String이며,
-// data는 {}의 객체입니다.
-// 팩토리 패턴을 통해 클래스명과 데이터를 받아서 올바른 타입으로 반환해줍니다.
-function createEntity(className, data) {
-  // className이 유효한지 확인합니다.
-  if (!Object.values(storageEnum.Class).includes(className))
-    return storageEnum.Result.Failure;
-
-  // data가 유효한 데이터인지 확인합니다.
-  if (!isValid(data)) return storageEnum.Result.Failure;
-
-  // 클래스반환을 위한 빈공간 생성
-  var instance = null;
-
-  // 클래스명을 통해 각각의 클래스를 반환합니다.
-  switch (className) {
-    case storageEnum.Class.Product: {
-      // 클래스의 여부를 체크합니다.
-      // 유효하지 않은 경우 바로 break 합니다.
-      if (Product.validate(data) === false) {
-        console.log("데이터 이상", data);
-        break;
-      }
-
-      // 문제가 없는 것으로 판단하여,
-      // 클래스를 생성자를 통해 생성해줍니다.
-      // tempClass에 생성한 것을 넣습니다.
-      // todo : tempClass대신에 instance로 바꾸기 helper : 소라님
-      const { name, price, image } = data;
-      // id는 클래스마다 다르므로 따라 구분하였습니다.
-      // const pnoId = data[getIdByClass[className]];
-      const pnoId = "";
-      instance = new Product(pnoId, name, price, image);
-      console.log(instance);
-      break;
-    }
-    case storageEnum.Class.Cart: {
-      // 클래스의 여부를 체크합니다.
-      // 유효하지 않은 경우 바로 break 합니다.
-      if (Cart.validate(data) === false) break;
-
-      // 문제가 없는 것으로 판단하여,
-      // 클래스를 생성자를 통해 생성해줍니다.
-      // tempClass에 생성한 것을 넣습니다.
-      // todo : tempClass대신에 instance로 바꾸기 helper : 소라님
-      const { product, model, quantity, select } = data;
-      instance = new Cart(product, model, quantity, select);
-      break;
-    }
-    case storageEnum.Class.User: {
-      // 클래스의 여부를 체크합니다.
-      // 유효하지 않은 경우 바로 break 합니다.
-      if (User.validate(data) === false) {
-        console.log("데이터 이상", data);
-        break;
-      }
-
-      // 문제가 없는 것으로 판단하여,
-      // 클래스를 생성자를 통해 생성해줍니다.
-      // tempClass에 생성한 것을 넣습니다.
-      // todo : tempClass대신에 instance로 바꾸기 helper : 소라님
-      const { id, pwd } = data;
-      // const pnoId = data[getIdByClass[className]];
-      const pnoId = "";
-      instance = new User(pnoId, id, pwd);
-      break;
-    }
-    case storageEnum.Class.Comment: {
-      // 클래스의 여부를 체크합니다.
-      // 유효하지 않은 경우 바로 break 합니다.
-      console.log(data);
-      if (Comment.validate(data) === false) break;
-
-      // 문제가 없는 것으로 판단하여,
-      // 클래스를 생성자를 통해 생성해줍니다.
-      // tempClass에 생성한 것을 넣습니다.
-      // todo : tempClass대신에 instance로 바꾸기 helper : 소라님
-      const { user, content, date } = data;
-      // const pnoId = data[getIdByClass[storageEnum.Class.Comment]];
-      const pnoId = "";
-      instance = new Comment(user, pnoId, content, date);
-      console.log("instance", instance);
-      break;
-    }
-    case storageEnum.Class.Review: {
-      // 클래스의 여부를 체크합니다.
-      // 유효하지 않은 경우 바로 break 합니다.
-      console.log("data", data);
-      console.log("coments", data.comments);
-      if (Review.validate(data) === false) break;
-
-      // 문제가 없는 것으로 판단하여,
-      // 클래스를 생성자를 통해 생성해줍니다.
-      // tempClass에 생성한 것을 넣습니다.
-      // todo : tempClass대신에 instance로 바꾸기 helper : 소라님
-
-      // const pnoId = data[getIdByClass[storageEnum.Class.Review]];
-      const pnoId = "";
-      const { product, user, title, content, date, comments } = data;
-      instance = new Review(
-        pnoId,
-        product,
-        user,
-        title,
-        content,
-        date,
-        comments
-      );
-      break;
-    }
-    case storageEnum.Class.QnA: {
-      // 클래스의 여부를 체크합니다.
-      // 유효하지 않은 경우 바로 break 합니다.
-      if (QnA.validate(data) === false) break;
-
-      const pnoId = "";
-      const { user, title, content, date } = data;
-      instance = new Review(pnoId, user, title, content, date);
-      break;
-    }
-  }
-
-  // 데이터에 문제가 없을 경우 데이터 저장, 문제가 있을 경우 실패를 반환
-  return instance !== null ? instance : storageEnum.Result.Failure;
-}
 
 // 데이터에 이상이 없는지 체크한다.
 // null이나 빈데이터, 혹은 이상한 데이터를 체크합니다.
 // 데이터 유효성 검사
-const isValid = (data) => {
-  // 널 , 언디파인 체크
-  if (data === null || data === undefined) return false;
 
-  // 스트링 체크
-  if (typeof data === "string" && data.trim() === "") return false;
-
-  // 숫자 체크
-  // Not a Number
-  if (typeof data === "number" && isNaN(data)) return false;
-
-  // 빈 배열 체크
-  if (Array.isArray(data) && data.length === 0) return false;
-
-  // 빈 객체 체크
-  // todo : 빈 객체 체크가 좀 난해하다. 배열 체크를 다시할 필요 있을까. : 소라님
-  if (typeof data === "object" && !Array.isArray(data)) {
-    if (Object.keys(data).length === 0) return false;
-  }
-
-  // 데이터가 문제 없다면 true를 반환합니다.
-  return true;
-};
+import {
+  createEntity,
+  isValid,
+  getIdByClass,
+  storageEnum,
+} from "./factories/factoriesApi";
 
 // 클래스 명과 데이터 배열을 매개 변수로 받아서,
 // 저장소의 초기 데이터를 세팅합니다.
